Bug Fix: Fixed wrong class types in TableFormats (BugId:  969, 971, 972, 979)  - Updated TableFormatTest

+++ src/main/java/net/nikr/eve/jeveasset/gui/tabs/orders/MarketTableFormat.java
@@ -124,7 +124,7 @@ public enum MarketTableFormat implements EnumTableColumn<MyMarketOrder> {
 			return from.getOutbidCount();
 		}
 	},
-	OUTBID_DELTA(Long.class, GlazedLists.comparableComparator()) {
+	OUTBID_DELTA(Double.class, GlazedLists.comparableComparator()) {
 		@Override
 		public String getColumnName() {
 			return TabsOrders.get().columnOutbidDelta();
